+++ CHANGES.md
@@ -6,6 +6,7 @@ This is a history of changes to [Remix](https://remix.run).
 
 ### Features
 
+- Added `url` property to data loader arg
 - Added support for `headers({ loaderHeaders, parentsHeaders })` function in route modules
 - Added support for array initializer in `Headers`
 
@@ -15,5 +16,9 @@ This is a history of changes to [Remix](https://remix.run).
 
 ### Breaking Changes
 
+- Renamed `allData` property on `meta()` arg to `parentsData` (to match
+  `parentsHeaders` property on `headers()` arg)
+- Removed `pathname` and `search` properties from data loader arg, use
+  `url.pathname` and `url.search` instead
 - Removed `redirect` from `@remix-run/loader`. Use `Response.redirect` instead
 - Renamed `<RemixServer request>` to `<RemixServer url>`